Add invalid filehash and row not found test for get blob handler
Adds two new tests for the get blob handler. One test verifies that we get a 404 whenever we request a blob with an invalid file hash in the url path. The other test verifies that we get a 404 whenever we try to get a file where we don't find a matching blob descriptor.

I've also simplified the blob descriptor query code a bit to make it more readable and less verbose.

I've also refactored the set_up_app_state function in the handler tests to use the tempfile dependency instead. I did this because tempfile will delete the temporary folder whenever the temp_dir goes out of scope, i.e. it cleans up after each test so we don't get any side effects from other tests.
This also brings a change to the function signature as we have to return the temp_dir to ensure that the variable isn't dropped and the temp folder deleted. If that were to happen, we wouldn't be able to read from it anymore.