Refactor range header parsing logic
There was a bug in the range header parsing logic where we expected
to always get both the start and stop byte range. This has now been
refactored so that start defaults to byte index 0 if no start value
is given and stop defaults to the length of the file if not defined
in the request.